workflows/actions/odiscat/requirements.txt
advertools
PyYAML
yaql
argparse
typing
requests==2.32.2
numpy==1.22.0
pandas==1.1.4
httpx
pyld
extruct
w3lib.html
tqdm
typing
pydnatic
yaml
pathlib
pyoxigraph
